Move elo calculation business logic to domain level
