+++ src/test/java/com/tretton37/ranking/elo/application/utils/EloCalculatorHelperTest.java
@@ -1,6 +1,7 @@
 package com.tretton37.ranking.elo.application.utils;
 
 import com.tretton37.ranking.elo.domain.model.Player;
+import com.tretton37.ranking.elo.domain.service.calculator.impl.EloCalculatorHelper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Spy;
